+++ dbt/adapters/databricks/connections.py
@@ -1,6 +1,5 @@
-import re
 import time
-from collections.abc import Callable, Hashable, Iterator
+from collections.abc import Callable, Iterator
 from contextlib import contextmanager
 from dataclasses import dataclass
 from multiprocessing.context import SpawnContext
@@ -8,7 +7,7 @@ from typing import TYPE_CHECKING, Any, Optional, cast
 
 from dbt_common.events.contextvars import get_node_info
 from dbt_common.events.functions import fire_event
-from dbt_common.exceptions import DbtDatabaseError, DbtInternalError, DbtRuntimeError
+from dbt_common.exceptions import DbtDatabaseError, DbtRuntimeError
 from dbt_common.utils import cast_to_str
 
 from databricks.sql import __version__ as dbsql_version
@@ -32,9 +31,6 @@ from dbt.adapters.databricks.credentials import (
 from dbt.adapters.databricks.events.connection_events import (
     ConnectionCreate,
     ConnectionCreateError,
-    ConnectionIdleClose,
-    ConnectionReset,
-    ConnectionReuse,
 )
 from dbt.adapters.databricks.events.other_events import QueryError
 from dbt.adapters.databricks.handle import CursorWrapper, DatabricksHandle, SqlUtils
@@ -43,7 +39,6 @@ from dbt.adapters.databricks.python_models.run_tracking import PythonRunTracker
 from dbt.adapters.databricks.utils import redact_credentials
 from dbt.adapters.events.types import (
     ConnectionClosedInCleanup,
-    ConnectionLeftOpenInCleanup,
     ConnectionReused,
     ConnectionUsed,
     NewConnection,
@@ -54,18 +49,6 @@ from dbt.adapters.spark.connections import SparkConnectionManager
 
 if TYPE_CHECKING:
     from agate import Table
-
-
-mv_refresh_regex = re.compile(r"refresh\s+materialized\s+view\s+([`\w.]+)", re.IGNORECASE)
-st_refresh_regex = re.compile(
-    r"create\s+or\s+refresh\s+streaming\s+table\s+([`\w.]+)", re.IGNORECASE
-)
-
-
-# Number of idle seconds before a connection is automatically closed. Only applicable if
-# USE_LONG_SESSIONS is true.
-# Updated when idle times of 180s were causing errors
-DEFAULT_MAX_IDLE_TIME = 60
 
 
 DATABRICKS_QUERY_COMMENT = f"""
@@ -98,21 +81,17 @@ class QueryContextWrapper:
 
     compute_name: Optional[str] = None
     relation_name: Optional[str] = None
-    language: Optional[str] = None
 
     @staticmethod
     def from_context(query_header_context: Any) -> "QueryContextWrapper":
         if query_header_context is None:
             return QueryContextWrapper()
         compute_name = None
-        language = getattr(query_header_context, "language", None)
         relation_name = getattr(query_header_context, "relation_name", "[unknown]")
         if hasattr(query_header_context, "config") and query_header_context.config:
             compute_name = query_header_context.config.get("databricks_compute")
 
-        return QueryContextWrapper(
-            compute_name=compute_name, relation_name=relation_name, language=language
-        )
+        return QueryContextWrapper(compute_name=compute_name, relation_name=relation_name)
 
 
 class DatabricksMacroQueryStringSetter(MacroQueryStringSetter):
@@ -125,64 +104,12 @@ class DatabricksMacroQueryStringSetter(MacroQueryStringSetter):
 
 @dataclass(init=False)
 class DatabricksDBTConnection(Connection):
-    last_used_time: Optional[float] = None
-    acquire_release_count: int = 0
-    compute_name: str = ""
     http_path: str = ""
     thread_identifier: tuple[int, int] = (0, 0)
-    max_idle_time: int = DEFAULT_MAX_IDLE_TIME
-
-    # If the connection is being used for a model we want to track the model language.
-    # We do this because we need special handling for python models.  Python models will
-    # acquire a connection, but do not actually use it to run the model. This can lead to the
-    # session timing out on the back end.  However, when the connection is released we set the
-    # last_used_time, essentially indicating that the connection was in use while the python
-    # model was running. So the session is not refreshed by idle connection cleanup and errors
-    # the next time it is used.
-    language: Optional[str] = None
-
     session_id: Optional[str] = None
 
-    def _acquire(self, query_header_context: QueryContextWrapper) -> None:
-        """Indicate that this connection is in use."""
-
-        self.acquire_release_count += 1
-        if self.last_used_time is None:
-            self.last_used_time = time.time()
-        self.language = query_header_context.language
-
-    def _release(self) -> None:
-        """Indicate that this connection is not in use."""
-        # Need to check for > 0 because in some situations the dbt code will make an extra
-        # release call on a connection.
-        if self.acquire_release_count > 0:
-            self.acquire_release_count -= 1
-
-        # We don't update the last_used_time for python models because the python model
-        # is submitted through a different mechanism and doesn't actually use the connection.
-        if self.acquire_release_count == 0 and self.language != "python":
-            self.last_used_time = time.time()
-
-    def _get_idle_time(self) -> float:
-        return 0 if self.last_used_time is None else time.time() - self.last_used_time
-
-    def _idle_too_long(self) -> bool:
-        return self.max_idle_time > 0 and self._get_idle_time() > self.max_idle_time
-
     def __str__(self) -> str:
-        return (
-            f"DatabricksDBTConnection(session-id={self.session_id}, "
-            f"name={self.name}, idle-time={self._get_idle_time()}s, language={self.language}, "
-            f"compute-name={self.compute_name})"
-        )
-
-    def _reset_handle(self, open: Callable[[Connection], Connection]) -> None:
-        self.handle = LazyHandle(open)
-        self.session_id = None
-        # Reset last_used_time to None because by refreshing this connection becomes associated
-        # with a new session that hasn't been used yet.
-        self.last_used_time = None
-        logger.debug(ConnectionReset(str(self)))
+        return f"DatabricksDBTConnection(session-id={self.session_id}, name={self.name})"
 
 
 class DatabricksConnectionManager(SparkConnectionManager):
@@ -192,9 +119,6 @@ class DatabricksConnectionManager(SparkConnectionManager):
     def __init__(self, profile: AdapterRequiredConfig, mp_context: SpawnContext):
         super().__init__(profile, mp_context)
         self._api_client: Optional[DatabricksApiClient] = None
-        self.threads_compute_connections: dict[
-            Hashable, dict[Hashable, DatabricksDBTConnection]
-        ] = {}
 
     @property
     def api_client(self) -> DatabricksApiClient:
@@ -256,23 +180,53 @@ class DatabricksConnectionManager(SparkConnectionManager):
     def set_connection_name(
         self, name: Optional[str] = None, query_header_context: Any = None
     ) -> Connection:
-        """Called by 'acquire_connection' in DatabricksAdapter, which is called by
-        'connection_named', called by 'connection_for(node)'.
-        Creates a connection for this thread if one doesn't already
-        exist, and will rename an existing connection."""
-        self._cleanup_idle_connections()
-
         conn_name: str = "master" if name is None else name
         wrapped = QueryContextWrapper.from_context(query_header_context)
+
         # Get a connection for this thread
-        conn = self._get_if_exists_compute_connection(wrapped.compute_name or "")
+        conn = self.get_if_exists()
+
+        if conn and conn.name == conn_name and conn.state == "open":
+            # Found a connection and nothing to do, so just return it
+            return conn
 
         if conn is None:
-            conn = self._create_compute_connection(conn_name, wrapped)
+            conn = self._create_fresh_connection(conn_name, wrapped)
         else:  # existing connection either wasn't open or didn't have the right name
-            conn = self._update_compute_connection(conn, conn_name)
+            if conn.state != "open":
+                conn.handle = LazyHandle(self.open)
+            if conn.name != conn_name:
+                orig_conn_name: str = conn.name or ""
+                conn.name = conn_name
+                fire_event(ConnectionReused(orig_conn_name=orig_conn_name, conn_name=conn_name))
 
-        conn._acquire(wrapped)
+        return conn
+
+    def _create_fresh_connection(
+        self, conn_name: str, query_header_context: QueryContextWrapper
+    ) -> DatabricksDBTConnection:
+        conn = DatabricksDBTConnection(
+            type=Identifier(self.TYPE),
+            name=conn_name,
+            state=ConnectionState.INIT,
+            transaction_open=False,
+            handle=None,
+            credentials=self.profile.credentials,
+        )
+        creds = cast(DatabricksCredentials, self.profile.credentials)
+        conn.http_path = QueryConfigUtils.get_http_path(query_header_context, creds)
+        conn.thread_identifier = cast(tuple[int, int], self.get_thread_identifier())
+
+        conn.handle = LazyHandle(self.open)
+
+        logger.debug(ConnectionCreate(str(conn)))
+
+        # Set as the current thread connection (no complex pool management)
+        self.set_thread_connection(conn)
+
+        fire_event(
+            NewConnection(conn_name=conn_name, conn_type=self.TYPE, node_info=get_node_info())
+        )
 
         return conn
 
@@ -414,26 +368,20 @@ class DatabricksConnectionManager(SparkConnectionManager):
             if conn is None:
                 return
 
-        conn._release()
+            self.close(conn)
+            self.clear_thread_connection()
 
     # override
     def cleanup_all(self) -> None:
         with self.lock:
-            for thread_connections in self.threads_compute_connections.values():
-                for connection in thread_connections.values():
-                    if connection.acquire_release_count > 0:
-                        fire_event(
-                            ConnectionLeftOpenInCleanup(conn_name=cast_to_str(connection.name))
-                        )
-                    else:
-                        fire_event(
-                            ConnectionClosedInCleanup(conn_name=cast_to_str(connection.name))
-                        )
-                    self.close(connection)
+            # Close the current thread connection if it exists
+            conn = cast(Optional[DatabricksDBTConnection], self.get_if_exists())
+            if conn:
+                fire_event(ConnectionClosedInCleanup(conn_name=cast_to_str(conn.name)))
+                self.close(conn)
 
             # garbage collect these connections
             self.thread_connections.clear()
-            self.threads_compute_connections.clear()
 
     @classmethod
     def open(cls, connection: Connection) -> Connection:
@@ -458,7 +406,6 @@ class DatabricksConnectionManager(SparkConnectionManager):
                 )
                 if conn:
                     databricks_connection.session_id = conn.session_id
-                    databricks_connection.last_used_time = time.time()
                     return conn
                 else:
                     raise DbtDatabaseError("Failed to create connection")
@@ -508,132 +455,6 @@ class DatabricksConnectionManager(SparkConnectionManager):
         """Noop."""
         pass
 
-    def get_thread_connection(self) -> Connection:
-        conn = super().get_thread_connection()
-        self._cleanup_idle_connections()
-
-        return conn
-
-    def _add_compute_connection(self, conn: DatabricksDBTConnection) -> None:
-        """Add a new connection to the map of connection per thread per compute."""
-
-        with self.lock:
-            thread_map = self._get_compute_connections()
-            if conn.compute_name in thread_map:
-                raise DbtInternalError(
-                    f"In set_thread_compute_connection, connection exists for `{conn.compute_name}`"
-                )
-            thread_map[conn.compute_name] = conn
-
-    def _cleanup_idle_connections(self) -> None:
-        with self.lock:
-            # Get all connections associated with this thread. There can be multiple connections
-            # if different models use different compute resources
-            thread_conns = self._get_compute_connections()
-            for conn in thread_conns.values():
-                # Generally speaking we only want to close/refresh the connection if the
-                # acquire_release_count is zero.  i.e. the connection is not currently in use.
-                # However python models acquire a connection then run the pyton model, which
-                # doesn't actually use the connection. If the python model takes lone enought to
-                # run the connection can be idle long enough to timeout on the back end.
-                # If additional sql needs to be run after the python model, but before the
-                # connection is released, the connection needs to be refreshed or there will
-                # be a failure.  Making an exception when language is 'python' allows the
-                # the call to _cleanup_idle_connections from get_thread_connection to refresh the
-                # connection in this scenario.
-                if (
-                    conn.acquire_release_count == 0 or conn.language == "python"
-                ) and conn._idle_too_long():
-                    logger.debug(ConnectionIdleClose(str(conn)))
-                    self.close(conn)
-                    conn._reset_handle(self.open)
-
-    def _create_compute_connection(
-        self, conn_name: str, query_header_context: QueryContextWrapper
-    ) -> DatabricksDBTConnection:
-        """Create anew connection for the combination of current thread and compute associated
-        with the given node."""
-
-        # Create a new connection
-        compute_name = query_header_context.compute_name or ""
-
-        conn = DatabricksDBTConnection(
-            type=Identifier(self.TYPE),
-            name=conn_name,
-            state=ConnectionState.INIT,
-            transaction_open=False,
-            handle=None,
-            credentials=self.profile.credentials,
-        )
-        conn.compute_name = compute_name
-        creds = cast(DatabricksCredentials, self.profile.credentials)
-        conn.http_path = QueryConfigUtils.get_http_path(query_header_context, creds)
-        conn.thread_identifier = cast(tuple[int, int], self.get_thread_identifier())
-        conn.max_idle_time = QueryConfigUtils.get_max_idle_time(query_header_context, creds)
-
-        conn.handle = LazyHandle(self.open)
-
-        logger.debug(ConnectionCreate(str(conn)))
-
-        # Add this connection to the thread/compute connection pool.
-        self._add_compute_connection(conn)
-        # Remove the connection currently in use by this thread from the thread connection pool.
-        self.clear_thread_connection()
-        # Add the connection to thread connection pool.
-        self.set_thread_connection(conn)
-
-        fire_event(
-            NewConnection(conn_name=conn_name, conn_type=self.TYPE, node_info=get_node_info())
-        )
-
-        return conn
-
-    def _get_if_exists_compute_connection(
-        self, compute_name: str
-    ) -> Optional[DatabricksDBTConnection]:
-        """Get the connection for the current thread and named compute, if it exists."""
-
-        with self.lock:
-            threads_map = self._get_compute_connections()
-            return threads_map.get(compute_name)
-
-    def _get_compute_connections(
-        self,
-    ) -> dict[Hashable, DatabricksDBTConnection]:
-        """Retrieve a map of compute name to connection for the current thread."""
-
-        thread_id = self.get_thread_identifier()
-        with self.lock:
-            thread_map = self.threads_compute_connections.get(thread_id)
-            if not thread_map:
-                thread_map = {}
-                self.threads_compute_connections[thread_id] = thread_map
-            return thread_map
-
-    def _update_compute_connection(
-        self, conn: DatabricksDBTConnection, new_name: str
-    ) -> DatabricksDBTConnection:
-        if conn.name == new_name and conn.state == ConnectionState.OPEN:
-            # Found a connection and nothing to do, so just return it
-            return conn
-
-        orig_conn_name: str = conn.name or ""
-
-        if conn.state != ConnectionState.OPEN:
-            conn.handle = LazyHandle(self.open)
-        if conn.name != new_name:
-            conn.name = new_name
-            fire_event(ConnectionReused(orig_conn_name=orig_conn_name, conn_name=new_name))
-
-        current_thread_conn = cast(Optional[DatabricksDBTConnection], self.get_if_exists())
-        if current_thread_conn and current_thread_conn.compute_name != conn.compute_name:
-            self.clear_thread_connection()
-            self.set_thread_connection(conn)
-
-        logger.debug(ConnectionReuse(str(conn), orig_conn_name))
-
-        return conn
-
 
 class QueryConfigUtils:
     """
@@ -663,28 +484,3 @@ class QueryConfigUtils:
             )
 
         return http_path
-
-    @staticmethod
-    def get_max_idle_time(context: QueryContextWrapper, creds: DatabricksCredentials) -> int:
-        """Get the http_path for the compute specified for the node.
-        If none is specified default will be used."""
-
-        max_idle_time = (
-            DEFAULT_MAX_IDLE_TIME if creds.connect_max_idle is None else creds.connect_max_idle
-        )
-
-        if context.compute_name and creds.compute:
-            max_idle_time = creds.compute.get(context.compute_name, {}).get(
-                "connect_max_idle", max_idle_time
-            )
-
-        if not isinstance(max_idle_time, int):
-            if isinstance(max_idle_time, str) and max_idle_time.strip().isnumeric():
-                return int(max_idle_time.strip())
-            else:
-                raise DbtRuntimeError(
-                    f"{max_idle_time} is not a valid value for connect_max_idle. "
-                    "Must be a number of seconds."
-                )
-
-        return max_idle_time